pyproject.toml
[tool.poetry]
name = "tgbot-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "tgbot_test"}]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.0"
requests = "^2.28.2"
beautifulsoup4 = "^4.11.1"
fake-useragent = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
